ScanParam, MultiValueScanParam, SingleValueScanParam moved to core/param.py
